updated services form

+++ saviour-billing-portal/src/components/clients/ClientList.jsx
@@ -2,9 +2,11 @@ import React, { useEffect, useState } from "react"
 import { deleteClient, listClient } from "../../services/ClientService"
 import { useNavigate } from "react-router-dom"
 
-const ClientList = ({ onLogout }) => {
+const ClientList = () => {
 
     const [clients, setClients] = useState([]);
+    const [errors, setErrors] = useState('');
+    const [loading, setLoading] = useState(true);
 
     const navigate = useNavigate();
 
@@ -12,11 +14,12 @@ const ClientList = ({ onLogout }) => {
 
     function getAllClient() {
         listClient().then(response => {
-            console.log("API Response:", response.data);
             setClients(response.data);
-
+            setLoading(false);
         }).catch(error => {
             console.error(error);
+            setErrors('Failed to load Clients. Please try again later.');
+            setLoading(false);
         });
     };
     const addNewClient = () => navigate('/add-client');
@@ -26,20 +29,29 @@ const ClientList = ({ onLogout }) => {
     }
 
     function removeClient(id) {
-        console.log(id);
-
-        deleteClient(id).then((response) => {
-            getAllClient();
-        }).catch(error => {
-            console.error(error);
-        });
+        if (window.confirm('Are you sure you want to delete this service?')) {
+            deleteClient(id).then((response) => {
+                getAllClient();
+            }).catch(error => {
+                console.error(error);
+            });
+        }
     };
+
+    if (loading) {
+        return <div className="text-center">Loading...</div>;
+    }
+
+    if (errors) {
+        return <div className="text-center text-danger">{errors}</div>;
+    }
+
     return (
         <div className="bg-white rounded">
             <div className="page-header">
                 <h5 className=' p-4 mb-0'>List Client</h5>
             </div>
-            <div className="page-body p-4">
+            <div className="page-body p-4 h-100">
                 <div className="table-container rounded">
                     <div className="table-header py-3"></div>
                     <table className="table mb-5 sticky-header">
@@ -59,8 +71,8 @@ const ClientList = ({ onLogout }) => {
                                         <td>{client.name}</td>
                                         <td>{client.email}</td>
                                         <td>
-                                            <button className="btn" onClick={() => updateClient(client.id)}><i className="bi bi-pencil-fill text-primary"></i></button>
-                                            <button className="btn"  onClick={() => removeClient(client.id)}><i className="bi bi-trash3-fill text-danger fw-bold"></i></button>
+                                            <button className="btn action-btn" onClick={() => updateClient(client.id)}><i className="bi bi-pencil-fill text-primary"></i></button>
+                                            <button className="btn action-btn" onClick={() => removeClient(client.id)}><i className="bi bi-trash3-fill text-danger fw-bold"></i></button>
                                         </td>
                                     </tr>
                                 ))
